[TODO] 완료: EXIT 타입 정의, 퇴장과 메시지 작성 response 구분하기
